src/requirements.txt
discord.py
requests-html
bs4
jikan
jikanpy
nekos.py
topggpy
lxml
retry
DateTime
validators
pygicord
pygelbooru
rule34Py
AnilistPython
validator
NHentai-API==0.0.16
discord-components
discord-webhook
pysaucenao
hmtai
